// src/types/flashcads.ts
export interface Deck {
  title: string; // Название карточки
  category: string; // Категория карточки (например, "Химия")
  cards: number; // Количество карточек
  students: number; // Количество студентов
  shares: number; // Количество поделившихся карточкой
}